经过大量实验，现在百分百肯定，这个2860芯片，撑不住我们内网的数据包 现在独立给它一个路由器，小型内网只有硬网关和我的台式机，非常干净 一次硬件错误都没有 硬件读写是2字节 取消定时输出日志

+++ Enc28j60.h
@@ -36,8 +36,8 @@ public:
     byte ReadReg(byte addr);
 	// 写寄存器值 发送写寄存器命令和地址
     void WriteReg(byte addr, byte data);
-    uint PhyRead(byte addr);
-    bool PhyWrite(byte addr, uint data);
+    ushort PhyRead(byte addr);
+    bool PhyWrite(byte addr, ushort data);
     void ClockOut(byte clock);
 	bool Linked();
